pyproject.toml
[tool.black]
line_length = "79"

[tool.isort]
profile = "black"
line_length = "79"

[tool.pylint.settings]
errors-only = "True"
disable = "E0401,E0611"
max-line-length = "79"

[tool.mypy]
ignore_missing_imports = true
scripts_are_modules = true
allow_untyped_decorators = true
strict = true

[[tool.mypy.overrides]]
module = ["tests.*"]
ignore_errors = true

[tool.poetry]
name = "latex_detector"
version = "0.1.0"
description = ""
authors = ["aleksandr <aleksandr_drobnov@epam.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
minio = "^7.1.6"
mmcv-full = {url = "https://download.openmmlab.com/mmcv/dist/1.2.1/torch1.7.0/cpu/mmcv_full-1.2.1%2Btorch1.7.0%2Bcpu-cp38-cp38-manylinux1_x86_64.whl"}
mmdet = "2.7.0"
torch = "1.7.0"
torchvision = "0.8.1"
scipy = "1.5.4"
pdfplumber = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "22.3.0"
isort = "5.9.1"
mypy = "0.910"
pylint = "2.8.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
